requirements.txt
# Web Framework
streamlit>=1.29.0

# LLM & AI APIs
google-generativeai>=0.3.0  # Google Gemini API

# PDF / DOCX processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.9.0
python-docx
PyPDF2
pymupdf 
pdfminer.six 
langdetect 
arabic-reshaper 
python-bidi
json

# OCR (optional, for scanned PDFs)
paddleocr>=2.6.0
easyocr>=1.6.0
# doctr>=0.6.0  # Comment out if installation issues

# Text processing / embeddings
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
orjson>=3.9.0  # Fast JSON parsing for RAG model
joblib>=1.3.0  # Model serialization for RAG indices

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.0.0

# Translation dependencies
googletrans==4.0.0rc1
deep-translator>=1.11.0

# Additional dependencies for contract processing
textract-py3>=1.5.0
python-magic-bin>=0.4.14  # For Windows file type detection
